sandbox/cgroup: move freeze/thaw code
The introduction of sandbox/cgroup package has allowed some unification
and cleanup of the cgroup code. The remaining element was the freezer
operation used by snap-update-ns.

This patch moves the freezer code to the sandbox/cgroup package.
Existing freeze/thaw functions became public, as did the mocking helper
to control freezing. Minimal documentation for the affected functions
was added as well.
