+++ sandbox/cgroup/export_test.go
@@ -18,6 +18,10 @@
  */
 package cgroup
 
+import (
+	. "gopkg.in/check.v1"
+)
+
 var (
 	Cgroup2SuperMagic  = cgroup2SuperMagic
 	ProbeCgroupVersion = probeCgroupVersion
@@ -37,4 +41,16 @@ func MockFsRootPath(p string) (restore func()) {
 	return func() {
 		rootPath = old
 	}
+}
+
+func MockFreezerCgroupDir(c *C) (restore func()) {
+	old := freezerCgroupDir
+	freezerCgroupDir = c.MkDir()
+	return func() {
+		freezerCgroupDir = old
+	}
+}
+
+func FreezerCgroupDir() string {
+	return freezerCgroupDir
 }
